+++ include/cpd/figtree.hpp
@@ -33,7 +33,7 @@ namespace cpd
 // Assume that the input matrices have already been tranposed so
 // callers can re-use transposed matrices.
 inline void figtree_wrap(arma::mat& Xt, arma::mat& Yt, arma::vec& q,
-                         double h, double epsilon, arma::vec& out, int eval_method = FIGTREE_EVAL_AUTO)
+                         double h, float epsilon, arma::vec& out, int eval_method = FIGTREE_EVAL_AUTO)
 {
     const arma::uword N = Xt.n_cols;
     const arma::uword M = Yt.n_cols;
